Refactor Manga model and update UI elements
This update entails notable changes to the Manga model in the application. More context around the 'status' of each manga work was required, hence the term 'status' is renamed to 'WorkState'. Additionally, more data is parsed from the source by the 'Scrapper' including the number of French volumes 'vfTomesCount' and the status of the French edition 'vfState'.

Updates are accompanied by changes in the UI to reflect these new states. The 'WorkDetailContent' now includes a 'ContentCell' component that adapts to the 'WorkType'. For Manga type, a 'StateCellManga' is rendered showing the states and number of volumes in both French and Original versions.

This commit is driven by the need to provide more details about the works to the users and adapt the UI for a better user experience.