Refine UI aesthetics in SearchContent view
This change refines the aesthetics in SearchContent view of the application by changing the mislocated padding in the 'WorkImage' layout. The padding was previously wrongly applied to the wrapper layout, causing irregularities in UI baseline. The padding has been correctly applied to 'Row' (a child layout) for accurate view spacing. Also, the extra bullet in the volumes count string creating redundancy has been removed for a cleaner representation in the string resources file.