Update UI and functionality related to Jmedia
Removed dependency on kotlin-bom from build.gradle as it is not necessary anymore. Added padding to several elements and improved the action buttons in HomeScreen.kt and WorkDetailScreen.kt for better user experience. Updated JScaledContent's press effect and added some animation for more natural response. For WorkDetailViewModel, added a handler for 'like' action and created a mutable state variable to track the like status.