Add voState and voTomesCount to Manga Model and adjust Manga Detail View
The additional voState and voTomesCount properties were added to the Manga.kt data model to store the state of the "version originale" and the tomes counts. Also, two commented out fields were removed. The WorkDetailScreen.kt was modified to display these new properties. The work's genres list was included in the work detail view. LazyRow was used to render the genres, each wrapped in a styled box. Adjustments were also made to the layout to make the view neater, such as, adding padding, divider and rearranging the elements. Modifications like these enhance the UX by providing an organised view of a Manga's details.