Remove Manga specific details from WorkDetailContent
The specific details related to displaying Manga genre has been removed from the WorkDetailContent file. This includes the removal of components like GenresListManga, StateCellManga and GenresMangaListItem. The change was implemented to enhance the code modularity and keep WorkDetailContent focused on high level details. The Manga specific details will be handled separately in MangaCell.