Update SearchView and WorksList layout and add navigation in SearchView
The SearchView's layout has been updated with a more intuitive and interactive UI. The `fetchWorks()` function now clears search text field focus after execution. A back navigation icon has been added to the TopBar. Also, the layout for displaying work details in the WorksListItem of the SearchContent view has been enhanced. An action button to mark a list item as favorite is added, and new strings have been defined in the resources for "empty description" and "research".