Add new heart plus icon and adjust UI on manga list
A new drawable resource "heart_plus_24px.xml" has been added to the project. This new icon is used in the "WorksListItem" function in the "SearchContent.kt" file replacing the previously used 'Favorite' icon. Additionally, the action buttons size has been increased to 150 dp and an extra 'LibraryAdd' icon was added with a divider for better UI separation.